Fix broken example
My experimental changes type errors in some of the examples in the
README. In particular, the output is no longer allowed to be an array of
arrays, because I don't have a standard shell representations for them.
Please use an array of tuples instead.

Now that my fork got more stars than the original, I should at least
keep the README up-to-date.