magic!
The default mode is now to infer the mode which should be used based on
the type.

This is backwards-compatible with the previous default of evaluating the
expression, because the type of those expressions were value types, not
function types, and the mode we use for those types is to evaluate the
expression.