workaround for resurected ghc bug?
Using the doctest library currently causes the tests to fail with the
following message:

  reference: Too late for parseStaticFlags: call it before newSession

There used to be an issue with that error message in an old version of
ghc [1], not sure why we are triggering it again with a recent version.
Regardless, the workaround is easy: use the executable version of
doctest instead of the library version.

[1] https://github.com/sol/doctest-haskell/issues/11