document and unify Config.Cache
As I was documenting Config.Cache, I discovered and fixed a small
inconsistency. If the goal of (<//>) was to avoid the awkwardness of
the code listing the components in the opposite order from what we are
constructing, why were other functions allowed to use the wrong order?
Fixed.

The doctest will fail on Windows, oh well.