more succinct help
I switched to the help generated by Hawk.Spec.Options, which has the
capability of generating exactly the same text as before, but I took the
opportunity to shorten the text so that most descriptions fit in one
line. I also changed "[options]" to "[option]..." to be closer to grep's
"[OPTION]...", which I assume must be standard, although I have also
seen "[OPTION ...]".