uniform indentation
I like to indent with two spaces, but whenever there is the possibility
of a where clause, I like to use four spaces in order to match the style
in [1]. I don't follow everything in [1], though.

I haven't changed the files which don't already follow this convention;
I don't want to impose my conventions everywhere, I just want to make
sure that each file is consistent with its own conventions.

[1] https://github.com/tibbe/haskell-style-guide/blob/master/haskell-style.md#indentation