split off runtime, fix crazy hClose bug
Having the source for the runtime and executable at the same place was
causing issues during the test. By giving the path to the executable
source, ghci (called by doctest) was also finding the source to the
runtime, which caused HawkRuntime values to be interpreted instead of
compiled. This in turn conflicted with hint, who is only using compiled
values except inside the user expression. That caused ghci to segfault,
and doctest to report an unexpectedly-closed handle.