slightly shorted Extra-Source-Files via wildcards
According to the cabal documentation, "The reason for providing only a
very limited form of wildcard is to concisely express the common case of
a large number of related files of the same file type without making it
too easy to accidentally include unwanted files". I still find it very
annoying that I can't just say "include everything in src recursively".