switch to cabal 1.18
We have had trouble in the past running tests with cabal 1.18,
especially with stringsearch, but now we are having trouble installing
stringsearch and haskell-src-exts with ghc 7.8 and cabal 1.16.

On my machine, running ghc 7.8.2, those packages install just fine with
cabal 1.18 and the tests also pass. Maybe it's time to accept cabal 1.18
back into our heart.