new version bound measurements
The following packages no longer impose a lower version bound, because
the algorithm from which those bounds were calculated was incorrect.

  bytestring
  directory
  filepath
  process

Those packages (and many others) are installed by ghc-7.6.3, and the
algorithm decided that the minimum version was the version installed
with ghc. That's probably not true. I haven't tried with older versions
of ghc, but if I did, then the algorithm would probably conclude that
the minimum version is the version installed by that version of ghc.

For the other packages, I re-ran the algorithm, and somehow the lower
bound for some of the packages has decreased! The code has changed a lot
since last time I ran cabal-rangefinder, so perhaps we are indeed using
fewer functions which are only available in modern packages. But it
could just as well be another bug in the algorithm.