make the Lock tests pass on Windows
for some reason the sleep lengths need to be multiplied by 100 in order
for the Lock tests to pass on Windows.

I don't think this has any effect on the correctness of the Lock itself,
which doesn't use any sleeps of any length during its normal operation;
it's just that the error messages are slightly different if socket
operations are used at a slightly incorrect time, versus a very
incorrect time.