switch to hpack
I used to use a custom Setup.hs which made sure that the library wasn't
just built before the tests, it was also installed, so that hint could
find it. This doesn't seem to be required anymore, as stack configures
an environment variable which gives hint access to the locally built
library.