delete the last failing System.Console.Hawk.Test
The failing test was that [] should evaluate to "", but instead has an
ambiguous type. Well, I agree that [] has an ambiguous type. Maybe older
versions of ghc were defaulting it to type [()]; but if newer versions
of ghc aren't, no big deal, I don't think this is a feature worth
keeping.