restrict when the job runs, not when to checkout
I cargo-culted the if from https://kodimensional.dev/github-actions, but
now my run-every-month run failed because that if returned false and the
checkout didn't run. reading the documentation more closely, I now
realize that the if is intended to restrict when CI should run, but is
instead restricting when the checkout runs. as a result, when CI is
triggered in circumstances which cause this if to be false, CI runs and
fails instead of not running.