rewrite the package-db detection code
instead of looking for all the environment variables and folders which
could be relevant and then failing if none are available, first detect
which installation method was used, then look at the place which that
specific installation method uses. otherwise complain that an
unsupported installation method was used.

also, we're now looking at the ghc version which was used to compile the
program in order to compute some of the file paths, so we should no
longer fail if there are package dbs for more than one version of ghc.