temporarily disable CI cache
CI failed with @Prelude.chr: bad argument@, which usually means that one
version of ghc is trying to load a file compiled with another version of
ghc. CI is compiling the code with multiple versions of ghc... but it's
supposed to be using separate caches!

Let's disable the cache to clarify whether caching is involved or not.